pyproject.toml
[project]
name = "coin-trader"
version = "0.1.0"
description = "Multi-agent cryptocurrency trading system for Upbit"
requires-python = ">=3.12"
dependencies = [
    "redis[hiredis]>=5.0",
    "httpx>=0.27",
    "websockets>=13.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "PyJWT>=2.8",
    "pandas>=2.2",
    "numpy>=1.26",
    "pandas-ta>=0.3.14b1",
    "scikit-learn>=1.5",
    "structlog>=24.1",
    "typer>=0.12",
    "rich>=13.7",
]

[project.scripts]
coin-trader = "coin_trader.__main__:app"

[project.optional-dependencies]
test = [
    "pytest>=8.2",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "respx>=0.22",
    "fakeredis[lua]>=2.23",
    "freezegun>=1.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["src/coin_trader"]
